+++ mobile/android/android-components/components/concept/toolbar/src/main/java/mozilla/components/concept/toolbar/Toolbar.kt
@@ -126,6 +126,14 @@ interface Toolbar {
     fun removePageAction(action: Action)
 
     /**
+     * Removes a previously added navigation action (see [addNavigationAction]). If the the provided
+     * actions was never added, this method has no effect.
+     *
+     * @param action the action to remove.
+     */
+    fun removeNavigationAction(action: Action)
+
+    /**
      * Declare that the actions (navigation actions, browser actions, page actions) have changed and
      * should be updated if needed.
      */