+++ mobile/android/android-components/components/feature/customtabs/src/test/java/mozilla/components/feature/customtabs/feature/CustomTabSessionTitleObserverTest.kt
@@ -87,6 +87,7 @@ class CustomTabSessionTitleObserverTest {
         override fun addPageAction(action: Toolbar.Action) = Unit
         override fun removePageAction(action: Toolbar.Action) = Unit
         override fun addNavigationAction(action: Toolbar.Action) = Unit
+        override fun removeNavigationAction(action: Toolbar.Action) = Unit
         override fun addEditAction(action: Toolbar.Action) = Unit
         override fun setOnEditListener(listener: Toolbar.OnEditListener) = Unit
         override fun displayMode() = Unit