+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarInteractorTest.kt
@@ -73,6 +73,10 @@ class ToolbarInteractorTest {
             fail()
         }
 
+        override fun removeNavigationAction(action: Toolbar.Action) {
+            fail()
+        }
+
         override fun setOnEditListener(listener: Toolbar.OnEditListener) {
             fail()
         }